requirements.txt
aiohttp==3.8.2
redis
gunicorn
python-decouple